+++ mrv2/lib/mrvCore/mrvHotkey.cpp
@@ -154,7 +154,7 @@ namespace mrv
     Hotkey kSOPSatNodes(false, false, false, false, 0);
 
     Hotkey kCopyFrameXYValues(true, false, false, true, 'c');
-    Hotkey kCopyRGBAValues(true, false, false, false, 'c');
+    Hotkey kCopyRGBAValues(true, false, true, false, 'c');
 
     Hotkey kSetInPoint(false, false, false, false, 'i');
     Hotkey kSetOutPoint(false, false, false, false, 'o');
@@ -191,6 +191,11 @@ namespace mrv
     Hotkey kToggleEditMode(false, false, false, false, 'e');
     Hotkey kToggleTimelineThumbnails(false, false, false, false, 0);
     Hotkey kToggleTimelineTransitions(false, false, false, false, 0);
+
+    Hotkey kEditCutFrame(true, false, false, false, 'x');
+    Hotkey kEditCopyFrame(true, false, false, false, 'c');
+    Hotkey kEditPasteFrame(true, false, false, false, 'v');
+    Hotkey kEditInsertFrame(true, false, false, false, 'i');
 
     Hotkey kRotatePlus90;  //( false, false, false, false, '+' );
     Hotkey kRotateMinus90; //( false, false, false, false, '-' );
@@ -372,6 +377,10 @@ namespace mrv
         HotkeyEntry(_("Next Channel"), kNextChannel),
         HotkeyEntry(_("First Frame"), kFirstFrame),
         HotkeyEntry(_("Last Frame"), kLastFrame),
+        HotkeyEntry(_("Cut Frame"), kEditCutFrame, true),
+        HotkeyEntry(_("Copy Frame"), kEditCopyFrame, true),
+        HotkeyEntry(_("Paste Frame"), kEditPasteFrame, true),
+        HotkeyEntry(_("Insert Frame"), kEditInsertFrame, true),
         HotkeyEntry(_("Toggle Menu Bar"), kToggleMenuBar),
         HotkeyEntry(_("Toggle Top Bar"), kToggleTopBar),
         HotkeyEntry(_("Toggle Pixel Bar"), kTogglePixelBar),