+++ mrv2/lib/mrvFl/mrvMenus.cpp
@@ -895,10 +895,26 @@ namespace mrv
                     _("Image/Version/Next"), kNextVersionImage.hotkey(),
                     (Fl_Callback*)next_image_version_cb, ui);
             }
-        }
 
-        menu->add(_("Edit/Slice"), 0, (Fl_Callback*)slice_clip_cb, ui);
-        menu->add(_("Edit/Remove"), 0, (Fl_Callback*)remove_clip_cb, ui);
+            menu->add(
+                _("Edit/Frame/Cut"), kEditCutFrame.hotkey(),
+                (Fl_Callback*)edit_cut_frame_cb, ui);
+            menu->add(
+                _("Edit/Frame/Copy"), kEditCopyFrame.hotkey(),
+                (Fl_Callback*)edit_copy_frame_cb, ui);
+            menu->add(
+                _("Edit/Frame/Paste"), kEditPasteFrame.hotkey(),
+                (Fl_Callback*)edit_paste_frame_cb, ui);
+            menu->add(
+                _("Edit/Frame/Insert"), kEditInsertFrame.hotkey(),
+                (Fl_Callback*)edit_insert_frame_cb, ui);
+            // menu->add(
+            //     _("Edit/Frame/Insert"), 0,
+            //     (Fl_Callback*)edit_insert_frame_cb, ui);
+
+            // menu->add(_("Edit/Slice"), 0, (Fl_Callback*)slice_clip_cb, ui);
+            // menu->add(_("Edit/Remove"), 0, (Fl_Callback*)remove_clip_cb, ui);
+        }
 
         // if ( num > 0 )
         // {